requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.34.2
apscheduler==3.10.4
pydantic==2.9.2
reportlab==4.2.2
